Time: 82 ms (30.87%) | Memory: 8.2 MB (41.61%) - LeetSync

+++ 1800-concatenation-of-consecutive-binary-numbers/concatenation-of-consecutive-binary-numbers.cpp
@@ -3,8 +3,8 @@
 class Solution {
 public:
     int concatenatedBinary(int n) {
-        long long int ans = 1;
-        long long int res = 2;
+        long int ans = 1;
+        long int res = 2;
         for(int i=2;i<n+1;i++)
         {
             ans = ans << res;
